requirements.txt
Flask>=1.1.2
Pillow==9.5.0
fontTools>=4.38.0
